+++ packages/runtime/src/roles.service.ts
@@ -103,14 +103,20 @@ export class RolesService {
     get effectiveRoles(): RoleState[] {
         let config = this._configuration;
 
-        if (config.mode === 'default')
-            return this._roles.filter(x => x.enabledByDefault !== false);
-        else if (config.mode == 'all-except')
-            return this._roles.filter(x => !config.roles.includes(x.class));
-        else if (config.mode == 'default-except')
-            return this._roles.filter(x => x.enabledByDefault !== false).filter(x => !config.roles.includes(x.class));
-        else if (config.mode == 'only')
-            return this._roles.filter(x => config.roles.includes(x.class));
+        if (config.mode === 'default') {
+            return this._roles
+                .filter(x => x.enabledByDefault !== false);
+        } else if (config.mode == 'all-except') {
+            return this._roles
+                .filter(x => !config.roles.includes(x.class) && !config.roles.includes(x.identifier));
+        } else if (config.mode == 'default-except') {
+            return this._roles
+                .filter(x => x.enabledByDefault !== false)
+                .filter(x => !config.roles.includes(x.class) && !config.roles.includes(x.identifier));
+        } else if (config.mode == 'only') {
+            return this._roles
+                .filter(x => config.roles.includes(x.class) || config.roles.includes(x.identifier));
+        }
         
         return [];
     }
@@ -126,16 +132,24 @@ export class RolesService {
         if (!SUPPORTED_ROLE_MODES.includes(config.mode))
             throw new InvalidOperationError(`Role mode '${config.mode}' is not supported (supports 'all-except', 'only')`);
         
+        let missingRoles = config.roles.filter(x => !this.roles.find(y => y.identifier === x || y.class === x));
+        if (missingRoles.length > 0) {
+            throw new Error(`The following roles have not been defined: ${missingRoles.join(', ')}. Did you define roles in altOnStart() instead of altOnInit()?`);
+        }
+    
         this._configuration = config;
     }
 
     getForModule(roleModuleClass) {
-        let role = this._roles.find(x => x.class === roleModuleClass);
+        let roles = this._roles.filter(x => x.class === roleModuleClass);
 
-        if (!role)
+        if (roles.length === 0)
             throw new ArgumentError(`Role module class ${roleModuleClass.name} is not registered`);
 
-        return role;
+        if (roles.length > 0)
+            throw new ArgumentError(`More than one role associated with module '${roleModuleClass.name}'`);
+
+        return roles[0];
     }
 
     getById(id : string) {